Infinite loop fix when message is "Connected to WS. Waiting for data images..."
During sudden price change, Bitmex MM tends to unreliably connected to websocket with the message "Connected to WS. Waiting for data images...". While MM message says that it is subscribed, WS symbol or account data was actually not downloaded causing an infinite loop. This commit fix that and exit the MM.